(PC-28532)[PRO] feat: edition preview for bookable offer
