+++ pro/src/components/CollectiveOfferNavigation/CollectiveOfferNavigation.module.scss
@@ -1,0 +1,19 @@
+@use "styles/mixins/_rem.scss" as rem;
+
+.duplicate-offer {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: rem.torem(8px);
+    margin-bottom: rem.torem(32px);
+}
+
+.tabs {
+    margin-top: rem.torem(24px);
+    margin-bottom: unset;
+
+    &-active {
+        margin-bottom: rem.torem(24px);
+    }
+}
+
