+++ pro/src/pages/AdageIframe/app/components/OfferInfos/AdageOffer/AdageOffer.module.scss
@@ -65,6 +65,7 @@
     display: flex;
     flex-direction: column;
     gap: rem.torem(16px);
+    word-break: break-word;
 
     &-item {
       &-subtitle {
