+++ pro/src/pages/AdageIframe/app/components/OffersInstantSearch/OffersSearch/Offers/PrebookingButton/PrebookingButton.tsx
@@ -23,7 +23,7 @@ const PrebookingButton = ({
   isInSuggestions,
   children,
   hideLimitDate,
-  isPreview,
+  isPreview = false,
 }: {
   className?: string
   stock: OfferStockResponse
