+++ pro/src/screens/CollectiveOfferSummaryEdition/CollectiveOfferSummaryEdition.module.scss
@@ -5,11 +5,3 @@
   top: 0;
   right: 0;
 }
-
-.duplicate-offer {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: rem.torem(8px);
-  margin-bottom: rem.torem(32px);
-}
