+++ pro/src/screens/CollectiveOfferSummaryEdition/CollectiveOfferSummaryEdition.tsx
@@ -1,6 +1,3 @@
-import React from 'react'
-import { useNavigate } from 'react-router-dom'
-
 import {
   GetCollectiveOfferTemplateResponseModel,
   GetCollectiveOfferResponseModel,
@@ -8,24 +5,10 @@ import {
 import ActionsBarSticky from 'components/ActionsBarSticky'
 import CollectiveOfferSummary from 'components/CollectiveOfferSummary'
 import OfferEducationalActions from 'components/OfferEducationalActions'
-import {
-  Events,
-  OFFER_FROM_TEMPLATE_ENTRIES,
-} from 'core/FirebaseEvents/constants'
-import {
-  createOfferFromTemplate,
-  isCollectiveOfferTemplate,
-  Mode,
-} from 'core/OfferEducational'
+import { isCollectiveOfferTemplate, Mode } from 'core/OfferEducational'
 import { computeURLCollectiveOfferId } from 'core/OfferEducational/utils/computeURLCollectiveOfferId'
 import { computeCollectiveOffersUrl } from 'core/Offers/utils'
-import useActiveFeature from 'hooks/useActiveFeature'
-import useAnalytics from 'hooks/useAnalytics'
-import useNotification from 'hooks/useNotification'
-import fullEditIcon from 'icons/full-edit.svg'
-import fullMoreIcon from 'icons/full-more.svg'
-import fullShowIcon from 'icons/full-show.svg'
-import { Button, ButtonLink } from 'ui-kit'
+import { ButtonLink } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
 
 import styles from './CollectiveOfferSummaryEdition.module.scss'
@@ -43,10 +26,6 @@ const CollectiveOfferSummaryEdition = ({
   reloadCollectiveOffer,
   mode,
 }: CollectiveOfferSummaryEditionProps) => {
-  const notify = useNotification()
-  const navigate = useNavigate()
-  const isMarseilleActive = useActiveFeature('WIP_ENABLE_MARSEILLE')
-
   const offerEditLink = `/offre/${computeURLCollectiveOfferId(
     offer.id,
     offer.isTemplate
@@ -57,13 +36,7 @@ const CollectiveOfferSummaryEdition = ({
     offer.isTemplate
   )}/collectif/stocks/edition`
 
-  const previewLink = `/offre/${computeURLCollectiveOfferId(
-    offer.id,
-    offer.isTemplate
-  )}/collectif/preview`
-
   const visibilityEditLink = `/offre/${offer.id}/collectif/visibilite/edition`
-  const { logEvent } = useAnalytics()
 
   return (
     <>
@@ -78,48 +51,6 @@ const CollectiveOfferSummaryEdition = ({
         reloadCollectiveOffer={reloadCollectiveOffer}
         mode={mode}
       />
-
-      {offer.isTemplate && (
-        <div className={styles['duplicate-offer']}>
-          <ButtonLink
-            link={{
-              to: offerEditLink,
-              isExternal: false,
-            }}
-            icon={fullEditIcon}
-          >
-            Modifier l’offre
-          </ButtonLink>
-          <ButtonLink
-            link={{
-              to: previewLink,
-              isExternal: false,
-            }}
-            icon={fullShowIcon}
-          >
-            Aperçu dans ADAGE
-          </ButtonLink>
-          <Button
-            variant={ButtonVariant.TERNARY}
-            icon={fullMoreIcon}
-            onClick={() => {
-              logEvent?.(Events.CLICKED_DUPLICATE_TEMPLATE_OFFER, {
-                from: OFFER_FROM_TEMPLATE_ENTRIES.OFFER_TEMPLATE_RECAP,
-              })
-              // eslint-disable-next-line @typescript-eslint/no-floating-promises
-              createOfferFromTemplate(
-                navigate,
-                notify,
-                offer.id,
-                undefined,
-                isMarseilleActive
-              )
-            }}
-          >
-            Créer une offre réservable
-          </Button>
-        </div>
-      )}
 
       <CollectiveOfferSummary
         offer={offer}
